Fix build of task_executor_test_type and cpu_device_test_type
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/130964
Tested-by: Sachkov, Alexey <alexey.sachkov@intel.com>

+++ opencl-intel/tests/task_executor_test_type/TaskExecutorTester.h
@@ -145,7 +145,7 @@ public:
     // Optimize By
     TASK_PRIORITY         GetPriority() const { return TASK_PRIORITY_MEDIUM;}
     TASK_SET_OPTIMIZATION OptimizeBy() const { return TASK_SET_OPTIMIZE_DEFAULT; }
-    unsigned int          PreferredSequentialItemsPerThread() const { return 1; }
+    size_t                PreferredSequentialItemsPerThread() const { return 1; }
 
 private:
